In-game menus working, main menu needs to be fixed
So close! The in-game menus are all working great with the bookmark. When you try to use them from the main menu, however, it all goes wrong.

I believe the solution is to transfer the functionality out of the Navigation screen and into the main menu itself. Then, ideally, the main menu will look like it does now, and the other menus will use the bookmark.