GUI UPDATE COMPLETE.
I finished off the main menu changes that were needed. Now all of the menus are working properly!!

There are still some polish updates to make: A Credits page. The End screen. Opening Title Page. Improving the backgrounds and the character portraits. But all of the basics of the GUI update are there.