Updated wildlife ambience
Updated the musical ambience with a loooong wildlife ambience track that I hope will make the sound less repetitive.

Cleaned up the game and packaged it up. I am now happy with the form of the game for 2021 :) STAGE 1: COMPLETE.

Avoiding committing wildlife2.wav because of the size of the file.